🚀 Implementing Roseau tool benchmark

+++ src/main/java/com/github/maracas/roseau/combinatorial/v2/benchmark/ToolResult.java
@@ -1,3 +1,3 @@
 package com.github.maracas.roseau.combinatorial.v2.benchmark;
 
-public record ToolResult(float executionTime, boolean isBreaking) {}
+public record ToolResult(long executionTime, boolean isBreaking) {}
